services-slack/requirements.txt
Django>=3.0.5,<3.1
mysqlclient
requests>=2.23.0,<2.24
slackclient>=2.5.0,<2.6
